Simplify migrator a bit
Instead of building up blobsToMigrate as an array, process them
immediately when finding out they don't exist in the destination.

This means not being able to know the number of blobs that aren't yet
migrated, but I think this isn't an issue. rsync tells you the *total*
number of files that are to be transferred, not just the ones that it
skips because the checksums match. This change results in showing a
progress bar for the total number of files that need to be transferred,
which remains the same each time unless the actual files in the source
grow in number.

This change also removes a buffered channel that had a size the length
of the blobs to migrate, which could be quite large.