Remove race condition
It's possible for the goroutine that's adding things to the
blobsToMigrate channel to put something on the channel and then close it
before a consumer from the channel can remove it. The consumer wouldn't
know either way; the channel would just be closed.

Adding the WaitGroup ensures we don't close the channel until all blobs
have been pulled and migrated.