Add and use BucketIterator
BucketIterator is a new interface with a concrete implementation for
each blobstore type, S3 and NFS.

BucketIterator is a replacement for Blobstore.List(). It only reads one
bucket at a time, and does not do checksumming. It does not build up an
array of all the items in the bucket, which means it can proceed to
uploading much more quickly than Blobstore.List().

I'm not happy with the current implementation of the constructors for
these iterators: I wanted the interface to be as easy as creating one
and then using Next() on it. This means that a lot of setup is done in
the constructors.