Move printing of status into watcher
With this change the blobstore migrator doesn't have to worry about
collecting errors or printing anything. It just tells the watcher what's
happening. The watcher does the printing and whatever else.

It also has the benefit of making the tests quieter, since the fake
watcher doesn't print stuff. If it was desired, we could also now test
that the right things are happening by testing calls to the watcher, but
I'll leave that out for now.