Remove test for bucket existence
Every call to the S3 blobstore Write method does an API call to list the
buckets to make sure the bucket you're writing to exists. If the bucket
does not exist it attempts to create the bucket.

This change removes both of these functionalities. If the bucket does
not exist it will return an error when trying to write to the bucket.

This speeds up the migrator, perhaps considerably depending on how many
buckets are in the S3 store. It also reduces the scope of the
permissions that the configured S3 key requires.